+++ R/rdf_has_bdb.R
@@ -13,3 +13,30 @@
 rdf_has_bdb <- function(){
  rdf_storage("BDB", new_db = TRUE, check_only = TRUE)
 }
+
+rdf_has_virtuoso <- function(user="dba", 
+                             password="dba", 
+                             dsn="Local Virtuoso"){
+#  has_driver <- rdf_storage("virtuoso", 
+#                            user = user, 
+#                            password = password, 
+#                            dsn = dsn, 
+#                            check_only=TRUE)
+  r <- tryCatch(rdf("virtuoso", user = user, 
+           password = password, dsn = dsn, fallback = FALSE),
+           error = function(e) FALSE)
+  if(is.logical(r)){
+    has_connection <- r
+  } else {
+    rdf_add(r, "", "dc:name", "bob")
+    if(length(r) >= 1){
+      has_connection <- TRUE
+    } else { 
+      has_connection <- FALSE
+    }
+    rdf_free(r)
+  }
+
+  has_connection
+  
+}